requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
openai
tiktoken
chromadb
pymupdf
docx2txt
python-docx
beautifulsoup4
lxml
pandas
requests
faiss-cpu==1.8.0.post4
